+++ src/Provider/Hook_Event_Provider.php
@@ -9,6 +9,7 @@ use Rarst\Laps\Event\Core_Events;
 use Rarst\Laps\Event\Genesis_Events;
 use Rarst\Laps\Event\Hybrid_Events;
 use Rarst\Laps\Event\THA_Events;
+use Rarst\Laps\Event\WooCommerce_Events;
 use Rarst\Laps\Event\Yoast_Events;
 
 /**
@@ -23,11 +24,12 @@ class Hook_Event_Provider implements ServiceProviderInterface {
 
 		$pimple['hook.events'] = function (): array {
 			return [
-				'core'    => new Core_Events(),
-				'genesis' => new Genesis_Events(),
-				'hybrid'  => new Hybrid_Events(),
-				'tha'     => new THA_Events(),
-				'yoast'   => new Yoast_Events(),
+				'core'        => new Core_Events(),
+				'genesis'     => new Genesis_Events(),
+				'hybrid'      => new Hybrid_Events(),
+				'tha'         => new THA_Events(),
+				'woocommerce' => new WooCommerce_Events(),
+				'yoast'       => new Yoast_Events(),
 			];
 		};
 	}